Moved some GPIO stuff to its own library
