+++ CMakeLists.txt
@@ -9,4 +9,4 @@ project(lit5v C CXX ASM)
 pico_sdk_init()
 
 add_subdirectory(src)
-add_subdirectory(include)
+
