+++ src/lit5v.c
@@ -2,47 +2,12 @@
  * Main for LIT-5V
  */
 
-// TODO: All these should be in an include
-
-// Output PWM signals on pins 0 and 1
-
 #include <stdio.h>
 #include "pico/stdlib.h"
-#include "hardware/gpio.h"
-#include "hardware/pwm.h"
-#include "hardware/adc.h"
 
-//// Inputs
-// Tonearm Positioning
-#define POS_REST_PIN 18
-#define POS_RECORD_PIN 19
-#define POS_STOP_PIN 22
-// Tonearm Tracking
-#define TRACK_IN_PIN 28
-#define TRACK_OUT_PIN 27
-#define TRACK_IN_ADC 2
-#define TRACK_OUT_ADC 1
-// Tonearm lift
-#define LIFT_UP_PIN 14
-#define LIFT_DOWN_PIN 15
+#include "lit5v_gpio.h"
 
-//// Outputs
-// Tonearm Transport
-#define TRANS_FWD_PIN 9
-#define TRANS_REV_PIN 8
-
-// Tonearm Lift
-#define LIFT_FWD_PIN 10
-#define LIFT_REV_PIN 11
-
-// Turntable
-#define TT_FWD_PIN 16
-
-//// Buttons
-// Lift/cue
-#define CUE_PIN 2
-
-
+// TODO: Handle these state variables in a better way, probably some encapsulating class
 bool stop_pin_high = false;
 bool rest_pin_high = false;
 
@@ -50,29 +15,6 @@ bool up_pin_high = false;
 bool down_pin_high = false;
 
 bool cue_pin_high = false;
-
-uint32_t pwm_set_freq_duty(uint slice_num,
-                           uint chan,uint32_t f, int d)
-{
-    uint32_t clock = 125000000;
-    uint32_t divider16 = clock / f / 4096 + (clock % (f * 4096) != 0);
-    if (divider16 / 16 == 0)
-        divider16 = 16;
-    uint32_t wrap = clock * 16 / divider16 / f - 1;
-    pwm_set_clkdiv_int_frac(slice_num, divider16/16,
-                            divider16 & 0xF);
-    pwm_set_wrap(slice_num, wrap);
-    pwm_set_chan_level(slice_num, chan, wrap * d / 100);
-    return wrap;
-}
-
-void handle_pin_event(uint32_t events, bool* pin_bool)
-{
-    if (events & GPIO_IRQ_EDGE_RISE)
-        *pin_bool = true;
-    else if (events & GPIO_IRQ_EDGE_FALL)
-        *pin_bool = false;
-}
 
 // TODO: Debounce these inputs somehow
 void gpio_callback(uint gpio, uint32_t events)
@@ -100,52 +42,6 @@ void gpio_callback(uint gpio, uint32_t events)
     {
         handle_pin_event(events, &cue_pin_high);
     }
-}
-
-void set_motor_pwm(uint32_t duty_cycle, bool direction, uint fwd_pin, uint rev_pin)
-{
-    uint32_t fwd_dc = 0;
-    uint32_t rev_dc = 0;
-
-    uint forward_slice_num = pwm_gpio_to_slice_num(fwd_pin);
-    uint forward_channel = pwm_gpio_to_channel(fwd_pin);
-
-    uint reverse_slice_num = pwm_gpio_to_slice_num(rev_pin);
-    uint reverse_channel = pwm_gpio_to_channel(rev_pin);
-
-    if (direction)
-    {
-        fwd_dc = duty_cycle;
-        rev_dc = 0;
-    }
-    else
-    {
-        fwd_dc = 0;
-        rev_dc = duty_cycle;
-    }
-
-    // PWM freq for motor controls should be at or above audible range (20KHz)
-    pwm_set_freq_duty(forward_slice_num, forward_channel, 20000, fwd_dc);
-    pwm_set_freq_duty(reverse_slice_num, reverse_channel, 20000, rev_dc);
-}
-
-// Convenience function for tonearm transport motor
-void set_tonearm_transport_dc(uint32_t duty_cycle, bool direction)
-{
-    set_motor_pwm(duty_cycle, direction, TRANS_FWD_PIN, TRANS_REV_PIN);
-}
-
-// Convenience function for tonearm list motor
-void set_tonearm_lift_dc(uint32_t duty_cycle, bool direction)
-{
-    set_motor_pwm(duty_cycle, direction, LIFT_FWD_PIN, LIFT_REV_PIN);
-}
-
-void set_turntable_dc(uint32_t duty_cycle)
-{
-    uint forward_slice_num = pwm_gpio_to_slice_num(TT_FWD_PIN);
-    uint forward_channel = pwm_gpio_to_channel(TT_FWD_PIN);
-    pwm_set_freq_duty(forward_slice_num, forward_channel, 5, duty_cycle);
 }
 
 void init_turntable_gpio()
